Undo renames
