+++ setup/src/artwork_shard.ts
