+++ setup/src/split_artwork_shard.ts
